Fix duplicate label names
There is an edge case where a label can appear both as internal and
external in a TSDB tenant. This happens when we remote-write from our
meta stack to receivers and k8s_cluster and observe_ingest_region get
written to the TSDB. When querying this data, receivers attach both the internal and external labels
and create a duplicate.

This is problematic only during a rollout from the previous version and against
queries for our meta tenant. The correct behavior is to preserve the external label value.