Fix coroutine race
The in-process client uses a coroutine which can be controlled through
next and stop calls. These functions should not be called concurrently
from two different goroutines, otherwise there could be a race condition.

This commit adds a mutex guarding these two calls. This could introduce
a bit of overhead but there is no other way to resolve the problem.